Commit 4: refactor
